requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.5
Flask-SQLAlchemy==3.0.0
